END-711: Fix register-leave-admin command (#3192)
https://lwd.atlassian.net/browse/END-711

This change fixes the register-leave-admin command. The command was passing incorrect metadata during registration, and was not passing metadata during verification of the user. While I was in there, I also added a flow to register-leave-admin to allow the password of a leave admin to be reset if it is run against an existing leave admin account. This should help us reset passwords more quickly in case that is needed (it is right now).

* END-711: Fix register-leave-admin command

* Fix CS

* Rethrow exception if it's not one we handle