API-1362 (1/?): Mark more integration tests (#3070)
Not strictly required, but could be considered loosely related to ticket, so
throwing it in here. Having things tagged appropriately may be useful to measure
shifts in the breakdown of unit vs integration tests.

Some light refactoring in `test_decode_cognito_token.py` for tests that could
easily be unit tests. Remove unnecessary `oauth_` fixtures and related test case.

Document the current usage/expectations (to what degree they exist). If we end
up wanting to make this more important, we can likely automate at least some of
this tagging with more work.

Add make targets running or excluding tests marked as integration, neither of
which are used for anything in our CI, but `test-unit` is convenient for "run
all tests that will be 'quick'", which may be useful.

Ticket: https://lwd.atlassian.net/browse/API-1362