INFRA-25 Follow-up Patch Lambda Errors (#3213)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-25

🛠 Changes
adjusted the datatype in the final output
Move BLOCK method
ℹ️ Context for reviewers
KDF had a lot of processing failed errors.

Example from the Test environment
image

This was narrowed down to be a function code error in the scrub_ip_addresses lambda code. Specifically, in the final data output:

data_record = {
    'recordId': record['recordId'],
    'result': 'Ok',
    'data': base64.b64encode(bytes(json.dumps(payload), 'utf-8'))
}
the data key was returning its value as a bytes string instead of a string which caused parsing errors.