END-719: "Successful Submission" notification test (#3270)
https://lwd.atlassian.net/browse/END-719

This PR adds a test for the "Successful Submission" notification, which is triggered when a claim is completely submitted. This notification shares a FINEOS trigger with the "Employer Request for Information" notification.
The test retrieves the email from testmail.app and checks that the correct NTN number is in the body of the email. The approval test, with the new notification test within it, was passig as of 8:40am EST on the Stage portal environment.

This PR also adds a longer timeout to the check for notifications and fixes an issue where the timestamp was being passed to the getEmails command as undefined.

* fix undefined timestamp issue

Add timeout

* add successful submit notif test to approval flow