INFRA-302: Log warning for missing employee/employer in eligibility check (#3235)
This is a valid scenario we expect, for cases where a claimant's TIN/SSN or employer FEIN is not found in the DOR data and they need to apply manually through the call center/fineos. logger.error calls are captured by Sentry so this should be converted to a warning.