END: Refactor generation code to support more advanced cases (#3242)
The existing employer/employee/claim generation code is complicated and cumbersome. In particular:

Needing to generate claims just to generate new DOR files is non-intuitive, and leads to a lot of extra data.
The lack of an "employee" object has made it impossible to do things that require specific employee info. Eg: see our extensive workarounds for supporting ID-verifiable employees in simulate.ts.
The generation code is primarily contained in one mega file, and is impossible to split up easily.
Data generation routines involve complex CLI commands, which can't be committed to version control or reviewed.
It's currently impossible to generate N claims of any particular type. You can generate "close to N", but not exactly that number. This is annoying since every other team expects us to deliver N cases.
To address these issues, and to make the next round of data generation tasks easier, I propose that data generation routines going forward are done with a single .ts file for each task, and that that .ts file uses a sane API to generate the data. This way, we'll be able to review and rerun these routines where necessary.

I've mocked up an example of what one of these scripts might look like for the upcoming training data load (note: This particular script isn't finalized - it's just an example).

Unfortunately, this PR is very big and far-reaching into the E2E code, since data generation is something we use for every aspect of E2E life. That being said, I think we should focus on 3 questions when reviewing this work:

Do Cypress tests pass at the same rate as before?
Are the data generation folks able to use it?
Does the code look "good enough" that we can live with it for now and adjust as needed?
If those three boxes are checked, I think we can get this in.

* Initial commit of employer/employee/claim generation refactor

* Cleanup

* Rename script

* WIP

* Add new submission handlers, pick employees out of the pool

* Add EmployeePool.slice()

* Add test for EmployerPool.pick() logic

* Cleanup

* Add scenario property back onto claim

* Start conversion of Cypress code

* Update typing for Cypress

* Drop backward compat for DocumentWaiter

* Update API methods after merge with main

* Remove unused variables

* Update scenarios so Cypress tests pass

* Fix generateCredentials

* Add scenarios

* Switch integration tests over to new generator

* Make generateClaim() compatible with new data generation tool

* Finalize and document submission process for scripts

* Allow claims to have arbitrary metadata attached

* Remove duplicate hydrate() call

* Fix direct to API file upload capabilities

* Fix code style

* Fix iterable.ts

* Fix tests

* Updating documentSize integration test to be more robust

* remove console.log()

Co-authored-by: Joseph Garcia <joe@lastcallmedia.com>