END-715: Generate fake employers that can be used to test LA verifications
The employer generator now set the DBA field to the employers name whereas previously the DBA field was set to and empty string. Employers quarterly withholding amounts are now generated programmatically based off the employers FEIN. When generating a pool of employer you can specify the amount of quarters you would like to set to zero.