INFRA-298 Refactor Slackbot Code to Stop Misreporting Failed Jobs (#3365)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-298

🛠 Changes
Terriyay code was refactored
Added the workflow-conclusion action which saves the final state of previous jobs (e.g. failed, skipped, successful) and rolls them into an environment variable.
ℹ️ Context for reviewers
When a particular job in a GHA workflow enters the ‘failed’ state, any future jobs that depend on that job do not also fail – they are instead assigned an end-state of skipped. This also applies for jobs that are cancelled: all future jobs that depend on a cancelled job are assigned the end-state of skipped.

The Slackbot code currently uses only the results of the api-release and gateway-release jobs (for the API), or just the invalidate-cloudfront job (for the Portal), to decide what the final state of a deployment should be. If one of those jobs fails, the Slackbot will report the whole deployment as failed. This is the intended behavior.