END: Make state tracker use atomic NDJSON writes (#3473)
END: Make state tracker use atomic NDJSON writes (#3473)

Converts the submission state tracker to use atomic appends when updating the state file. Also converts to using ndjson, which is necessary if we are writing a single line at a time. This is to fix a bug we discovered during the last large data submission where the state file was emptied after the process was cancelled at the wrong moment.