API: fix missing line breaks in output when running in docker (#3506)
## 🎫 Ticket

None

## 🛠 Changes

- Fix Makefile and decodelog to add missing line breaks in output.

## ℹ️ Context for reviewers

When running local commands in docker (the default), line breaks are missing in the output, making it difficult to read. This may not happen in all environments. The exact source of this problem is unclear but it appears related to terminal settings or stream redirection used by docker. I found the fix by experimenting.

## 👀 Demo

Before:
![Screenshot_2021-03-23_12-53-57](https://user-images.githubusercontent.com/3811269/112185924-0bf88d00-8bd7-11eb-97e3-af33cd6191d6.png)

After:
![Screenshot_2021-03-23_12-54-57](https://user-images.githubusercontent.com/3811269/112185952-1155d780-8bd7-11eb-9111-f8d4cfc95ae8.png)

## 🌡 End to End testing

N/A

Reviewers: @rocketnova @chouinar