API: move initialize_sentry() call in migrations/run.py (#3509)
## 🎫 Ticket

None

## 🛠 Changes

- Move call to `initialize_sentry()` from top level of `migrations/run.py` into `up()` function.

## ℹ️ Context for reviewers

This call was added in #3079, but after #3183 this file is imported from `db/__init__.py`, which is imported from almost all code. That means that the call to initialize sentry happens at import time for almost all components, which is unwanted. Instead move it into the `up()` function, which is the one that runs on releases.

## 👀 Demo

None

## 🌡 End to End testing

N/A

Reviewers: @aplybeah