INFRA-356: Pull E2E tests from the last Portal deployment branch (#3527)
https://lwd.atlassian.net/browse/INFRA-356

    When running E2E tests after a deployment, pull the version of E2E tests that exist on the related portal deploy branch.

ℹ️ Context for reviewers

The E2E tests are sometimes updated based on Portal UI/behavior changes, and may not run successfully for older Portal versions. Because of this, we need to be running the version of E2E tests that exist on the portal deploy branch.

e.g. if we deploy API stage, we should run the E2E tests from deploy/portal/stage.