END-751: Set feature flag for LST on new registration system (#3584)
https://lwd.atlassian.net/browse/END-751

Set a feature flag to enable the new registration system for LST and Cypress. Following the addition of this feature flag, all Portal registrations pass through the PFML API. See https://lwd.atlassian.net/browse/CP-1793 for more information on this change. We still have followup work to do in the E2E codebase to transition to using the PFML API rather than Cognito for registration, but that can follow (Cognito registration should continue to work for the time being).


* END-751: Set feature flag for LST on new registration system

* Also set feature flag for Cypress