END-756: Update flow for accessing UAT Fineos w/SSO creds (#3596)
This PR includes an updated flow for accessing Fineos UAT. Previously all that was needed for Fineos access was an username and password. With the recent change to UAT - you can only access UAT Fineos w/an authenticated Microsoft SSO account. This PR adds the needed flow for UAT access in order to run the e2e suite in this environment.

This PR also includes other small maintenance in regards to updating descriptions and email handling for the Fineos creating claim tests.