INFRA-324 Alert Slack when Auto-deploys to Test Fail (#3586)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-324

🛠 Changes
changed the conditional in the finish-comms job so that it sends a message when the status of the start-comms job is a failure, it will send a message
ℹ️ Context for reviewers
Currently we don't send any messages when autodeployments to test happen. This is to prevent noisyness.

However, we should know when they fail so we can remediate early on. For example, if the migration step fails in a test autodeploy, we should know about it.