END-762: Add message body wildcard match and use in all tests (#3661)
https://lwd.atlassian.net/browse/END-762

We've had a lot of trouble with false failures on notification emails (specifically the claimant-facing "application started" message, which doesn't even include a name in the subject line). This PR attempts to address that by adding a new way to search for messages that considers the message body content. It's now possible to limit to only e-mails that contain particular text (eg: claim #) in their HTML body.

We achieve this by testing the body of each e-mail that matches the filter specification, and requerying if the e-mail isn't found. This is following the suggestion of TestMail.

With this in place, we should stop receiving false failures - if a notifications test fails, it should be indicative of a notification that wasn't actually received, not another notification coming in at the same time.

Finally, I noticed that email.getNotificationData was unnecessarily returning a promise, which forced us deeper into callbacks. I switched it so it just returns a value directly, which helps with readability.

* END-762: Add message body wildcard match and use in all tests

* Remove duplicative not.be.empty check

* Remove async for e-mail parsing

* Lint fixes

* Fix approval/denial url checking