END: Enrich portal submission errors by adding error metadata (#3673)
We've recently seen a number of application submission failures that are extremely confusing to debug. The only way you'd know what's causing them is by knowing what this means: Cannot stash undefined data. Example.

The cause of these errors is failure during claim submission. This PR catches this kind of error and provides a much more helpful error message that includes any errors that the API returned, and a link to debug the failed request in New Relic.

* END: Enrich portal errors by adding error metadata

* Improve message