END-692: Create a test for checking the benefit amount - wage 90k (#3421)
🎫 Ticket
https://lwd.atlassian.net/browse/END-692

🛠 Changes
This will be a test for a scenario with a continuous medical leave and an employee that makes 90k per year and has a rotating 3x12 schedule.

This test checks the actual amount to be paid, which is the max benefit amount of $850 and is listed on Financials -> Payment History → Amounts Pending. It also adjusts the 30k test to check the actual amount.

In order to generate a consistent payment, the leave dates have been extended to two full weeks since the first week of leave is unpaid.

ℹ️ Context for reviewers
Testing instructions, background context, more in-depth details of the implementation, and anything else you'd like to call out or ask reviewers.

This is similar to an existing payments test named "Jill" which has a continuous bonding leave with a salary of 30k and a regular schedule.

This test submits a claim and employer response through the API using the e2e data set, adjudicates and approves the claim in Fineos, then checks the benefit amount in Fineos at Absence case > financials>Recurring payments> Benefit amount. This is the amount that Fineos is expected to put into the PEI extract.