PUB-130: Addresses live exclusively on the payment (#3638)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PUB-130

## 🛠 Changes

This modifies our address logic for PUB payments to make it so claimants can have multiple address pairs and payments have a single address pair. All downstream logic should be updated with this change as well.

We agreed with Mass that validating addresses associated with just a claimant prior to receiving the payment records was not needed and I have removed that from the address validation logic.

## ℹ️ Context for reviewers

It is possible (but should be uncommon) that a claimant can have two payments with differing addresses, or that a claimant switches to an address that they previously used. This allows us to handle those scenarios, much like we do for PUB EFT. For the extract logic, this follows a nearly identical pattern to the PUB EFT approach.