INFRA-381: Ignore push_db in New Relic API error alarm (#3671)
https://lwd.atlassian.net/browse/INFRA-381

🛠 Changes

    Ignore push_db in the New Relic API error alarm
    Require two consecutive windows of high error rate to trigger critical alarm

ℹ️ Context for reviewers

Another New Relic oddity. push_db is recorded as a separate transaction, and there's no good way to disable this. This is mentioned in one of the python-agent release notes.

This basically doubles the error rate since API calls are no longer being double-counted, so I'm also requiring two consecutive windows to be above 10% error rate in order for the alarm to trigger.