API-1522: Claimants added to FINEOS manually should be stored back to the employee tables (#3660)
If employee record sent from FINEOS is not found in PFML DB using the EMPLOYEEIDENTIFIER then search from employee using the SSN (EMPLOYEENATIONALID). If we cannot find it through either key add the employee to the PFML DB. These additions require us to use fields not used before from the extract: first name and last name. We have also added the updating of fineos_customer_number for completeness.

If the employee record FINEOS sent is found using the SSN we flag as a discrepancy (report.emp_id_discrepancies_count). Going forward these should be few as we would get the FINEOS employee adds over night and the DOR record probably takes a bit longer to arrive to our DB. These records have to be remediated using some sort of reconciliation between FINEOS and PFML DB as the employee_id cannot be modified in the PFML DB without affecting many other associated tables.

If the FINEOS extract does not provide an SSN for a record we cannot identify using the EMPLOYEEIDENTIFIER we also flag this as an error (report.no_ssn_present_count) and do not process the record.