END-680: Employer Verification Happy Path (#3711)
This PR includes an e2e test for leave admin self verification as well as a reusable portal flow for the steps necessary to verify a leave admin.

The test registers and logs in a new leave admin. Once the newly created leave admin is logged in they are given the option to verify their account. The test then goes through the process of successfully verifying the leave admin.

This verification unlike the verification done at account creation (email code). Once leave admins are verified for an organization they can then respond to leave requests against that organization. (After a 15 minute waiting period, so that FINEOS can register the user as a POC)