END: Update message when a test is skipped (#3713)
Adjusts the behavior when a test is skipped due to failed test dependencies. This takes us from sync skip; aborting execution to This test failed because it depends on a previous test that did not pass. It also marks the tests as failed rather than leaving them in "pending" status, which seems more accurate.