END: More E2E mail debugging improvements (#3725)
Some of our tests that fail during fetching e-mails are still failing with useless error messages. This is because the task they're attached to has a 60 second timeout, vs the 120 second timeout we use for the Testmail API calls. This PR brings those two things into alignment with each other by introducing a getEmails() function, which wraps the task and keeps the timeouts in sync with each other.

Also:

Adds missing debug information to the e-mail fetching done during Fineos claim submission tests.
Resolves two fatal issues with the ignored MHAP1 test.

* END: Add debug info to fineos case creation

* END: Set timeout of TestMail API calls

* Update submission type