INFRA-96 Run Tests in Parallel (#3618)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-96

🛠 Changes
added xdist
ℹ️ Context for reviewers
Our testing suite is expansive, but running a large number of tests is time consuming. Right now, our full test suite takes a long time to run, and as we add additional features and tests, that time only continues to increase (prior to merging this, tests average about 40-50 minutes).

How This Works:
Here is a more in-depth, under-the-hood glance at x-dist

As a very brief and watered down summary, there are two sets of nodes: the main/master1 and the workers. Each worker collects test-ids and then sends them back to the master node, which checks that all workers collected the same tests. Afterwords, the tests are distributed to each worker based on which algorithm is set.

1 this terminology is used interchangeably in xdist's docs, issues, etc. It can get confusing For simplicity, I'm going to use master as that is what's used in the linked overview.