END: Fix timeout on application started notification (#3740)
Adjusts the timeout on waiting for the application started notification. We've seen this notification take a little over a minute to arrive, even though it's the second notification received in this test. Eg.

This PR additionally tweaks the debug information for TestMail. The new debugging information should include livequery: false, and a timestamp_from property that matches when we stopped looking for e-mails. This should make debugging more accurate, as it will look at the window where we were actually looking for the e-mail.