END: Fix explicit dates not being respected when generating reduced and intermittent claims (#3746)
Specified dates were being overridden when using the claim generator to produce reduced and intermittent leave claims. The dates were ignored because there was no check to see if dates were provided in the claim spec before generating a random date range. This PR includes logic to check for explicit leave dates in the spec.