INFRA-376: Fix RDS log subscription to New Relic (#3662)
https://lwd.atlassian.net/browse/INFRA-376

🛠 Changes

    Fixes cloudwatch log group name to match the actual RDS log group
    Adds temporary deployment step to ensure no log groups are deleted

ℹ️ Context for reviewers

I was looking for RDS logs in new relic but they weren't there. Checked cloudwatch and there were two RDS log groups for each environment. The one with hyphens had logs and the one with underscores did not.

AWS creates these log groups automatically with the hyphens, so the ones in our terraform config need to match. Otherwise the new relic subscription is on the empty log group and nothing gets sent to NR.
Ensuring a smooth rollout

In order to ensure a smooth rollout, I added a temporary deployment step to import the hyphenated log group into the terraform state. If we didn't have this step, Terraform would try to create the hyphenated log group and would run into an error because the log group already exists.