CP-1948: Care for a Family Member enum (#3782)
🎫 Ticket
https://lwd.atlassian.net/browse/CP-1948

🛠 Changes
 Change all references of caring leave enum from Care For A Family Member to Care for a Family Member
 Add new leave reason qualifier enum Serious Health Condition

ℹ️ Context for reviewers
This Jira thread discussed the proper casing for the caring leave reason enum. Rather than Care For A Family Member, we should be using Care for a Family Member. Also discussed is the leave reason qualifier for caring leave, which is Serious Health Condition, a new enum added in this PR.