END-784: Generate and use employers specifically for LASR (#3779)
PFML-540: Implement GET claims endpoint (#145)

https://lwd.atlassian.net/browse/END-784

To fix the issue described in "Context for reviewers", this PR generates particular employers which will be used for the register_leave_admin test. Our regular employers will only have the single admin registered for them. Our specific register_leave_admin employers will have 0 employees, so it will be impossible to submit claims against them.

So this change:

Allows for metadata to be attached and used to query employers.
Updates the e2e employee/employer generation script to generate a couple of employers with metadata that makes them appropriate for use in LASR.
Updates the config to use this new data.
Bonus: Resolves a longstanding issue where the register-leave-admin command would hang infinitely. This was coming from the setTimeout call on our TestMailClient.

* END-784: Generate and use employers specifically for LASR

* Add new employers/employees, update config

* Add additional employer for LASR test

* Generate 2 LA employers

* Resolve hanging command after creating a leave admin

* Use Buffer.from instead of new buffer