Add an e2e command for running portal smoke tests (#3682)
Adds a new cypress:open:local:portal npm script so it's easier for portal engineers to smoke test portal changes locally. This way there's no need to edit the cypress.json file to get a reasonable subset of tests to run for a smoketest. Testing locally this subset of tests run in about 15 minutes. Unfortunately there are intermittent failures but it's a place to start.

* Add an e2e command for running portal smoke tests