END-788: Fix bug in portal test's Report Other leave & Benefits section (#3832)
This PR addresses a bug in our portal flow code for "Report leave & Other Benefits" section. This part of the portal flow contains similar selectors like clicking yes or no - then on the next page selecting yes or no again. What happened during e2e runs sporadically was the test moved to fast before the next page loaded (see example here) ... and because it's looking for a selector that is still on the page the entire flow is a step ahead and inevitably cause the test to fail.

This will be mitigated by asserting unique text for the intended page to be loaded before making any "yes" or "no" selections.

Co-authored-by: Joseph Garcia <joe@lastcallmedia.com>