API-1579: Strip commas from name fields in claimant list for DIA  (#3799)
DIA is comparing the names with fields that are stripped of commas anyway, so
just strip them out on our side instead of bailing.

Also split apart the `EmployeeFactory` to a separate base
`EmployeeOnlyDORDataFactory`. Was originally going to update the name fields to
include the occasional name with a comma to better match the real data seen in
production, but decided against that here. Splitting up the factory to be able
to generate Employee record with just the data that comes in via DOR could be
useful though, so keeping it (and it mirrors the structure for Employers).

Ticket: https://lwd.atlassian.net/browse/API-1579