API-981: DIA report (#3784)
- Support generating and sending DIA payment report. The sending part was mostly
  already in `reductions/reports/dia_payments_reports.py`, just moved to
  `reductions/reports/dia_payments/send.py` (and corresponding shifts with it's
  tests).
- Duplicate payment rows in the report to include info on each claim for the
  employee (the DIA half of https://lwd.atlassian.net/browse/API-1509).
- Some misc. CSV utilities to streamline some things

Both `dia_reduction_payment.fineos_customer_number` and
`employee.fineos_customer_number` are not indexed. This will likely result in
full scans of the `employee` table in order to generate the report (some 3.4
millions rows currently). The report generation time is not critical, but can
revisit if we need to add indexes if it turns out to be an issue.

Ticket: https://lwd.atlassian.net/browse/API-981