API-1624: Temporarily disable deps-check step in API CI (#4007)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-1624

## ℹ️ Context for reviewers

This action is ran on each pull request and is intended to block a PR from being merged if the action is failing. However, there's a [known dependency issue](https://lwd.atlassian.net/browse/API-1624) that is unable to be resolved immediately. Currently, people are ignoring this CI action and merging PRs with it failing. This is bad practice and could result in other actual issues being merged into the `main` branch since the `lint` and `format-check` steps aren't being ran.