CP-2034: use bonding leave plan proof in ApplicationCard (#4071)
🎫 Ticket
https://lwd.atlassian.net/browse/CP-2028
Subtask: https://lwd.atlassian.net/browse/CP-2034

🛠 Changes
For future bonding leave, the ApplicationCard checks for whether a certification document has previously been uploaded in order to show conditional content directing the user to upload the doc when it's available. Because of the switch to use the new plan proofs, the check for certification documents needed to be updated to include the right plan proof for bonding leave.