PFMLPB-1235: LST for employer dashboard (#4087)
In order to load test the employer dashboard, we should modify our current load tests by going to the employer dashboard in order to submit an employer response to a claim. In addition to navigating to the dashboard, we should test the pagination functionality. In this PR the load test will navigate to the second page of claims and look for the claim there. If the claim is not on the second page, we will navigate back to the first page and continue the employer response by selecting the claim from the first page of claims.

One thing I would like to improve upon is a hardcoded wait time in helpers.ts. Currently, there is a wait time of 8 seconds. Without waiting the test will run into failures and I've not been able to get the test to pass using Element's waitiers. I would be happy to pair to improve this if needed.