API-1590: Continue Using Old Doc Type As Needed (#4108)
For a period of time, the API needs to support both the old document type, State managed Paid Leave Confirmation, and the new types introduced by the new plan proofs. This PR implements code that allows the API to determine which document type to use. It also ensures that documents with the new document type are marked as evidence received when uploaded.