PFMLPB-1312: Switch LST code bundling to Webpack (#4155)
https://lwd.atlassian.net/browse/PFMLPB-1312

Switches to generating our Flood bundle using Webpack instead of simply zipping up a directory. There are many, many enhancements that follow from this work, but this just includes the initial change in bundler.

This PR also includes a new job to run on E2E code changes that asserts that the bundle produced for flood is actually executable. This should prevent us from unintentionally breaking the bundling process in the future.

* END: Implement webpack bundling of Flood files

* Replace Bundler with WebpackBundler

* Fix lint issue

* Fix lint error on dynamic import

* Add E2E job to check flood bundling

* Fix missing environment variable

* Fix broken yaml

* Add missing token

* Skip playwright download

* Remove SIMULATION_SPEED

* Pass additional environment variables to compilation

* Add missing FINEOS_USERS environment variable