PFMLPB-929: Track gender on Applications and send to CPS  (#3865)
This PR sets up tracking gender on applications and sending them to CPS. A gender is an optional string that can be on of five values, which each maps to a value of the "Gender" customer field in FINEOS. A column with the FINEOS gender terms was added to the FINEOS lookup table in the database and a gender ID column was added to the application table. When an application with a gender id is submitted to the API, the FINEOS customer gender should update as well. Applicants that are in progress when this feature is released (i.e., have completed the "Verify Identity" step but not completed Part One) should be able to submit their application without being prompted to enter Gender.

Co-authored-by: Melquisedeque Pereira <melqui1@live.com.pt>
Co-authored-by: Tyler Stephens <roberttstephens@gmail.com>
Co-authored-by: James Bursa <james@navapbc.com>