END: Fineos Service Pack Integration (#4193)
Updates document types and tasks to be used following the Fineos service pack. The new document types and caring leave test can be enabled by using the HAS_FINEOS_SP configuration setting in config.json. The plan is to roll this setting out one environment at a time, and as soon as the value is set to "true", the environment will be be tested against the new document types.

* WIP on integrated branch for DT3 testing

* Fix broken test

* Setup default value of HAS_FINEOS_SP

* Fix linting

* Fix JSON issue

* Disable caring leave flag if we're not on an SP environment

* Fix environment variables, gate caring leave test

* Fix linting issue