PUB-206: Payment extract and post processing puts errors in an additional  writeback state (#4243)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PUB-206

## 🛠 Changes

This modifies the payment extract and post-processing logic to put payments in a new PEI writeback state as well as setup the information for the writeback.

Additionally, the payment error state was split into two, the existing one, and a new restartable state.

## ℹ️ Context for reviewers

This required some reimagining of a few approaches. The way we determine active payments relied on states, but the states we had were no longer truly all restartable, some errors were non-restartable. So, the state was split into two.

In the post processing logic, I rewrote the way it finds prior payments to not look for "active" payments, but instead look for payments that have been sent to PUB (regardless of their current state).

To set a payment to PendingActive (the state we receive it in), we just make the Status column in the CSV blank.

See also: https://lucid.app/lucidchart/35eb55c3-d25b-496f-9bab-c7c9fdd3aedc/edit?page=qQjEcLYaGsxI#