PFMLPB-934: Add Gender to Portal Application Flow (#3862)
Ticket:
https://lwd.atlassian.net/browse/PFMLPB-934

Description:
This PR adds a portal page in the Verify Identity section which collects applicant gender. There are fives possible options: "Prefer not to answer", "Man", "Woman", "Non-binary", and "Gender not listed". "Prefer not to answer" is selected by default. It also adds gender to the Review page.

Demo:
https://www.loom.com/share/ae3b992e7fc4461885d0d952d370f127

Co-authored-by: Tyler Stephens <roberttstephens@gmail.com>
Co-authored-by: Melquisedeque Pereira <melqui1@live.com.pt>
Co-authored-by: Sawyer Hollenshead <git@sawyerh.com>