CP-2226: Add maintenance page for 6/6 (#4296)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-2226

## 🛠 Changes

* Adds a maintenance page for 6/6/2021 from 3:59am to 5:00am, ET

## ℹ️ Context for reviewers

I tested this locally by copying the config changes to my development environment and changing my computer's clock. We will test it again when it goes to stage.

The plan is for this to be included in v18.0 and to be deployed to Prod on June 2nd. That way the current maintenance window on 5/30 in v17.0 won't be overwritten.

## 👀 Demo

N/A