CP-2152: Add functionality to translate to/from fineos eform enums (#4303)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-2152

## 🛠 Changes

* Adds a new FineosToApiEnumConverter abstract class for translating FINEOS enum values to API enum values
* Adds a new `enumOverride` property to `PROP_MAP`s in both to_fineos and from_fineos transforms. This allows us to transform to/from FINEOS enums during the transform step

## ℹ️ Context for reviewers

Some of our enum values have diverged from FINEOS eform enum values, specifically PreviousLeaveQualifyingReasons and EmployerBenefitTypes. One approach to solve this is to go about updating our enum values to match the FINEOS enum values but this is difficult for a few reasons:
* Our enum values are used by the Portal and the API. Changing them requires coordinating the changes over multiple deployments or else breaking functionality during the deployment
* Our enum values are stored in the DB in a lookup table. In order to change them we have to use a DB migration. This would require _even more_ coordination of the changes over multiple deployments and carefully supporting the old and new values in the API code during the DB migration
* We have to support two diverged enums in the V1 and V2 other income eforms anyways -- changes to the V2 enums were made to help with the call center script

So instead in this PR I'm aiming to decouple FINEOS eform enum values from API enum values. This PR shows how it works with the PreviousLeaveQualifyingReason enum, which will be required to integrate with the Other Leave v2 eform.

This PR isn't attempting to integrate fully with the Other Leave v2 eform. That will happen in a separate PR. This PR just focuses on the enum transform logic.