PFMLPB-1285: Add new document for caring leave, update medical form (#4367)
This is some followup work in regard to adding "new" documentation for claim generation. Previously we updated the HCP forms (Serious Health Condition Form) to a new version and this PR adds a brand new document when generating caring leave claims.

Also fixes a small error with dates that get returned for caring leave that need to be in a specific window. The update there just makes sure the start date is the beginning of the week. W/o this errors occurred because the start date returned didn't cover a full 2 week period for checking payments.

Co-authored-by: Joseph Garcia <joe@lastcallmedia.com>