PFMLPB-1410: Add coverage for submitting medical pre-birth claims (#4365)
We've recently discovered that we do not have coverage for medical pre-birth scenarios while trying to submit QA claims for the 6/4 training load. This test will help the early identification of issues that may arise for pre-birth claim submission in the claimant portal.

Adds a Cypress test that submits a medical pre-birth claim through the portal.
Adds a medical pre-birth scenario.
Update the caring RFI scenario to avoid CI error checks.