API-1687: Call service agreement code only when an employer is created. (#4391)
Call Service Agreement code only when employers are created. Moved call to fineos_actions method from job to fineos_actions method where we determine whether to create or update an employer.

When an employer is updated the DOR import publishes an UPDATE event in the employer_log table. When the job to load employers reads these events it does not know what fields changed so it calls the service agreement and publishes a revision of the agreement even if the exemption fields have not changed. This is creating noise and processing issues in FINEOS.