END-676: Trigger LST baseline run on API perf deploys (#3190)
Triggers a small load test run on API deployments to the performance environment. This load test will be run for 30 minutes, at a concurrency that approximates production traffic (~6 claim submissions/minute). It will be triggered in parallel with the E2E test run, but since the load test service takes a while to fire up, we expect that these two things will not overlap.

The purpose of this trigger is currently data collection, not alerting. We would like to be able to assess performance on a more regular basis, keeping a record of when it deviated. It is expected that if performance dipped to a level where it was critical, New Relic alerts for the performance environment would tell us that. As such, this trigger just kicks off the job and does not wait around to see it complete.